// src/payload-types.ts
export interface User {
  id: string;
  /**
   * Görünen ad.
   */
  name?: string | null;
  /**
   * Admin: Tam yetki | Editor: İçerik yönetimi
   */
  role?: ('admin' | 'editor') | null;
  updatedAt: string;
  createdAt: string;
  email?: string | null;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}